Prevent solutions where rate < -1
Redefine the extension of the value in the range rate < 1 so that it
is not possible to get a zero there.

Also, handle the special case when the rate is -1 and years is zero by
returning the amount (i.e. if no time had gone by you still have your
money).  This prevents -1 from being a solution for all cases.